Moved auth filter to auth.js
